refactor(frontend): migrate to src layout and modularize chat components
- Migrate codebase structure to 'src' directory (src/app, src/components, src/lib)
- Update paths aliases in tsconfig.json to point to the src folder
- Extract state management and simulation actions into ChatContext
- Modularize UI pages into Sidebar, Chatroom, PersonalSettings, and GroupSettings components
- Configure dynamic routes for chatrooms (/chat/[chatId]) and settings (/settings) using Next.js App Router
- Clean up original monolithic app/page.tsx and app/layout.tsx files
