+++ frontend/src/app/(main)/layout.tsx
@@ -1,0 +1,32 @@
+"use client";
+
+import React from "react";
+import { ChatProvider, useChat } from "@/context/ChatContext";
+import Sidebar from "@/components/layout/Sidebar";
+
+function MainLayoutContent({ children }: { children: React.ReactNode }) {
+  const { isAuthenticated, isMounted } = useChat();
+
+  if (!isMounted || !isAuthenticated) {
+    return (
+      <div className="flex h-screen w-full items-center justify-center bg-background text-foreground font-sans">
+        載入中...
+      </div>
+    );
+  }
+
+  return (
+    <div className="flex h-screen w-full overflow-hidden bg-background text-foreground font-sans transition-colors">
+      <Sidebar />
+      {children}
+    </div>
+  );
+}
+
+export default function MainLayout({ children }: { children: React.ReactNode }) {
+  return (
+    <ChatProvider>
+      <MainLayoutContent>{children}</MainLayoutContent>
+    </ChatProvider>
+  );
+}
