+++ frontend/src/app/(main)/settings/page.tsx
@@ -1,0 +1,8 @@
+"use client";
+
+import React from "react";
+import PersonalSettings from "@/components/settings/PersonalSettings";
+
+export default function SettingsPage() {
+  return <PersonalSettings />;
+}
