+++ frontend/src/components/ui/Avatar.tsx
