+++ frontend/src/components/ui/ChatBubble.tsx
